feat (indexer): cleanup; rpc flags; pagination
