+++ go/indexer/db/db.go
@@ -3,303 +3,34 @@ package db
 import (
 	"database/sql"
 	"errors"
-	"math/big"
-
-	"github.com/google/uuid"
 
 	l2common "github.com/ethereum-optimism/optimism/l2geth/common"
 	"github.com/ethereum/go-ethereum/common"
+
+	// NOTE: Only postgresql backend is supported at the moment.
 	_ "github.com/lib/pq"
 )
 
-const createL1BlocksTable = `
-CREATE TABLE IF NOT EXISTS l1_blocks (
-	hash VARCHAR NOT NULL PRIMARY KEY,
-	parent_hash VARCHAR NOT NULL,
-	number INTEGER NOT NULL,
-	timestamp INTEGER NOT NULL
-)
-`
-
-const createL2BlocksTable = `
-CREATE TABLE IF NOT EXISTS l2_blocks (
-	hash VARCHAR NOT NULL PRIMARY KEY,
-	parent_hash VARCHAR NOT NULL,
-	number INTEGER NOT NULL,
-	timestamp INTEGER NOT NULL
-)
-`
-
-const createDepositsTable = `
-CREATE TABLE IF NOT EXISTS deposits (
-	guid VARCHAR PRIMARY KEY NOT NULL,
-	from_address VARCHAR NOT NULL,
-	to_address VARCHAR NOT NULL,
-	l1_token VARCHAR NOT NULL REFERENCES l1_tokens(address),
-	l2_token VARCHAR NOT NULL,
-	amount VARCHAR NOT NULL,
-	data BYTEA NOT NULL,
-	log_index INTEGER NOT NULL,
-	block_hash VARCHAR NOT NULL REFERENCES l1_blocks(hash),
-	tx_hash VARCHAR NOT NULL
-)
-`
-
-const createL1TokensTable = `
-CREATE TABLE IF NOT EXISTS l1_tokens (
-	address VARCHAR NOT NULL PRIMARY KEY,
-	name VARCHAR NOT NULL,
-	symbol VARCHAR NOT NULL,
-	decimals INTEGER NOT NULL
-)
-`
-
-const createL2TokensTable = `
-CREATE TABLE IF NOT EXISTS l2_tokens (
-	address TEXT NOT NULL PRIMARY KEY,
-	name TEXT NOT NULL,
-	symbol TEXT NOT NULL,
-	decimals INTEGER NOT NULL
-)
-`
-
-const createStateBatchesTable = `
-CREATE TABLE IF NOT EXISTS state_batches (
-	index INTEGER NOT NULL PRIMARY KEY,
-	root VARCHAR NOT NULL,
-	size INTEGER NOT NULL,
-	prev_total INTEGER NOT NULL,
-	extra_data BYTEA NOT NULL,
-	block_hash VARCHAR NOT NULL REFERENCES l1_blocks(hash)
-);
-CREATE INDEX IF NOT EXISTS state_batches_block_hash ON state_batches(block_hash);
-CREATE INDEX IF NOT EXISTS state_batches_size ON state_batches(size);
-CREATE INDEX IF NOT EXISTS state_batches_prev_total ON state_batches(prev_total);
-`
-
-const createWithdrawalsTable = `
-CREATE TABLE IF NOT EXISTS withdrawals (
-	guid VARCHAR PRIMARY KEY NOT NULL,
-	from_address VARCHAR NOT NULL,
-	to_address VARCHAR NOT NULL,
-	l1_token VARCHAR NOT NULL,
-	l2_token VARCHAR NOT NULL REFERENCES l2_tokens(address),
-	amount VARCHAR NOT NULL,
-	data BYTEA NOT NULL,
-	log_index INTEGER NOT NULL,
-	block_hash VARCHAR NOT NULL REFERENCES l2_blocks(hash),
-	tx_hash VARCHAR NOT NULL,
-	state_batch INTEGER REFERENCES state_batches(index)
-)
-`
-
-const insertETHL1Token = `
-INSERT INTO l1_tokens
-	(address, name, symbol, decimals)
-VALUES ('0x0000000000000000000000000000000000000000', 'Ethereum', 'ETH', 18)
-ON CONFLICT (address) DO NOTHING;
-`
-
-// earlier transactions used 0x0000000000000000000000000000000000000000 as
-// address of ETH so insert both that and
-// 0xDeadDeAddeAddEAddeadDEaDDEAdDeaDDeAD0000
-const insertETHL2Token = `
-INSERT INTO l2_tokens
-	(address, name, symbol, decimals)
-VALUES ('0xDeadDeAddeAddEAddeadDEaDDEAdDeaDDeAD0000', 'Ethereum', 'ETH', 18)
-ON CONFLICT (address) DO NOTHING;
-INSERT INTO l2_tokens
-	(address, name, symbol, decimals)
-VALUES ('0x0000000000000000000000000000000000000000', 'Ethereum', 'ETH', 18)
-ON CONFLICT (address) DO NOTHING;
-`
-
-const createL1L2NumberIndex = `
-CREATE UNIQUE INDEX IF NOT EXISTS l1_blocks_number ON l1_blocks(number);
-CREATE UNIQUE INDEX IF NOT EXISTS l2_blocks_number ON l2_blocks(number);
-`
-
-type PaginationParam struct {
-	Limit  uint64
-	Offset uint64
-}
-
-var schema = []string{
-	createL1BlocksTable,
-	createL2BlocksTable,
-	createL1TokensTable,
-	createL2TokensTable,
-	createStateBatchesTable,
-	insertETHL1Token,
-	insertETHL2Token,
-	createDepositsTable,
-	createWithdrawalsTable,
-	createL1L2NumberIndex,
-}
-
-type TxnEnqueuedEvent struct {
-	BlockNumber uint64
-	Timestamp   uint64
-	TxHash      common.Hash
-	Data        []byte
-}
-
-func (e TxnEnqueuedEvent) String() string {
-	return e.TxHash.String()
-}
-
-type Deposit struct {
-	GUID        string
-	TxHash      common.Hash
-	L1Token     common.Address
-	L2Token     common.Address
-	FromAddress common.Address
-	ToAddress   common.Address
-	Amount      *big.Int
-	Data        []byte
-	LogIndex    uint
-}
-
-func (d Deposit) String() string {
-	return d.TxHash.String()
-}
-
-type Withdrawal struct {
-	GUID        string
-	TxHash      l2common.Hash
-	L1Token     l2common.Address
-	L2Token     l2common.Address
-	FromAddress l2common.Address
-	ToAddress   l2common.Address
-	Amount      *big.Int
-	Data        []byte
-	LogIndex    uint
-}
-
-func (w Withdrawal) String() string {
-	return w.TxHash.String()
-}
-
-type IndexedL1Block struct {
-	Hash       common.Hash
-	ParentHash common.Hash
-	Number     uint64
-	Timestamp  uint64
-	Deposits   []Deposit
-}
-
-func (b IndexedL1Block) String() string {
-	return b.Hash.String()
-}
-
-type IndexedL2Block struct {
-	Hash        l2common.Hash
-	ParentHash  l2common.Hash
-	Number      uint64
-	Timestamp   uint64
-	Withdrawals []Withdrawal
-}
-
-func (b IndexedL2Block) String() string {
-	return b.Hash.String()
-}
-
-func (b *IndexedL1Block) Events() []TxnEnqueuedEvent {
-	nDeposits := len(b.Deposits)
-	if nDeposits == 0 {
-		return nil
-	}
-
-	var events = make([]TxnEnqueuedEvent, 0, nDeposits)
-	for _, deposit := range b.Deposits {
-		events = append(events, TxnEnqueuedEvent{
-			BlockNumber: b.Number,
-			Timestamp:   b.Timestamp,
-			TxHash:      deposit.TxHash,
-			Data:        deposit.Data, // TODO: copy?
-		})
-	}
-
-	return events
-}
-
-type StateBatch struct {
-	Index     *big.Int
-	Root      common.Hash
-	Size      *big.Int
-	PrevTotal *big.Int
-	ExtraData []byte
-	BlockHash common.Hash
-}
-
-type StateBatchJSON struct {
-	Index          uint64 `json:"index"`
-	Root           string `json:"root"`
-	Size           uint64 `json:"size"`
-	PrevTotal      uint64 `json:"prevTotal"`
-	ExtraData      []byte `json:"extraData"`
-	BlockHash      string `json:"blockHash"`
-	BlockNumber    uint64 `json:"blockNumber"`
-	BlockTimestamp uint64 `json:"blockTimestamp"`
-}
-
-type Token struct {
-	Address  string `json:"address"`
-	Name     string `json:"name"`
-	Symbol   string `json:"symbol"`
-	Decimals uint8  `json:"decimals"`
-}
-
-var ETHL1Token = &Token{
-	Address:  "0x0000000000000000000000000000000000000000",
-	Name:     "Ethereum",
-	Symbol:   "ETH",
-	Decimals: 18,
-}
-
-var ETHL2Address = l2common.HexToAddress("0xDeadDeAddeAddEAddeadDEaDDEAdDeaDDeAD0000")
-
-var ETHL2Token = &Token{
-	Address:  "0xDeadDeAddeAddEAddeadDEaDDEAdDeaDDeAD0000",
-	Name:     "Ethereum",
-	Symbol:   "ETH",
-	Decimals: 18,
-}
-
-type DepositJSON struct {
-	GUID           string `json:"guid"`
-	FromAddress    string `json:"from"`
-	ToAddress      string `json:"to"`
-	L1Token        *Token `json:"l1Token"`
-	L2Token        string `json:"l2Token"`
-	Amount         string `json:"amount"`
-	Data           []byte `json:"data"`
-	LogIndex       uint64 `json:"logIndex"`
-	BlockNumber    uint64 `json:"blockNumber"`
-	BlockTimestamp string `json:"blockTimestamp"`
-	TxHash         string `json:"transactionHash"`
-}
-
-type WithdrawalJSON struct {
-	GUID           string          `json:"guid"`
-	FromAddress    string          `json:"from"`
-	ToAddress      string          `json:"to"`
-	L1Token        string          `json:"l1Token"`
-	L2Token        *Token          `json:"l2Token"`
-	Amount         string          `json:"amount"`
-	Data           []byte          `json:"data"`
-	LogIndex       uint64          `json:"logIndex"`
-	BlockNumber    uint64          `json:"blockNumber"`
-	BlockTimestamp string          `json:"blockTimestamp"`
-	TxHash         string          `json:"transactionHash"`
-	Batch          *StateBatchJSON `json:"batch"`
-}
-
+// Database contains the database instance and the connection string.
 type Database struct {
 	db     *sql.DB
 	config string
 }
 
+// Close closes the database.
+// NOTE: "It is rarely necessary to close a DB."
+// See: https://pkg.go.dev/database/sql#Open
+func (d *Database) Close() error {
+	return d.db.Close()
+}
+
+// Config returns the db connection string.
+func (d *Database) Config() string {
+	return d.config
+}
+
+// GetL1TokenByAddress returns the ERC20 Token corresponding to the given
+// address on L1.
 func (d *Database) GetL1TokenByAddress(address string) (*Token, error) {
 	const selectL1TokenStatement = `
 	SELECT name, symbol, decimals FROM l1_tokens WHERE address = $1;
@@ -348,6 +79,8 @@ func (d *Database) GetL1TokenByAddress(address string) (*Token, error) {
 	return token, nil
 }
 
+// GetL2TokenByAddress returns the ERC20 Token corresponding to the given
+// address on L2.
 func (d *Database) GetL2TokenByAddress(address string) (*Token, error) {
 	const selectL2TokenStatement = `
 	SELECT name, symbol, decimals FROM l2_tokens WHERE address = $1;
@@ -396,6 +129,9 @@ func (d *Database) GetL2TokenByAddress(address string) (*Token, error) {
 	return token, nil
 }
 
+// AddL1Token inserts the Token details for the given address into the known L1
+// tokens database.
+// NOTE: a Token MUST have a unique address
 func (d *Database) AddL1Token(address string, token *Token) error {
 	const insertTokenStatement = `
 	INSERT INTO l1_tokens
@@ -424,6 +160,9 @@ func (d *Database) AddL1Token(address string, token *Token) error {
 	})
 }
 
+// AddL2Token inserts the Token details for the given address into the known L2
+// tokens database.
+// NOTE: a Token MUST have a unique address
 func (d *Database) AddL2Token(address string, token *Token) error {
 	const insertTokenStatement = `
 	INSERT INTO l2_tokens
@@ -452,6 +191,9 @@ func (d *Database) AddL2Token(address string, token *Token) error {
 	})
 }
 
+// AddIndexedL1Block inserts the indexed block i.e. the L1 block containing all
+// scanned Deposits into the known deposits database.
+// NOTE: the block hash MUST be unique
 func (d *Database) AddIndexedL1Block(block *IndexedL1Block) error {
 	const insertBlockStatement = `
 	INSERT INTO l1_blocks
@@ -513,6 +255,9 @@ func (d *Database) AddIndexedL1Block(block *IndexedL1Block) error {
 	})
 }
 
+// AddIndexedL2Block inserts the indexed block i.e. the L2 block containing all
+// scanned Withdrawals into the known withdrawals database.
+// NOTE: the block hash MUST be unique
 func (d *Database) AddIndexedL2Block(block *IndexedL2Block) error {
 	const insertBlockStatement = `
 	INSERT INTO l2_blocks
@@ -574,6 +319,8 @@ func (d *Database) AddIndexedL2Block(block *IndexedL2Block) error {
 	})
 }
 
+// AddStateBatch inserts the state batches into the known state batches
+// database.
 func (d *Database) AddStateBatch(batches []StateBatch) error {
 	const insertStateBatchStatement = `
 	INSERT INTO state_batches
@@ -606,7 +353,9 @@ func (d *Database) AddStateBatch(batches []StateBatch) error {
 	})
 }
 
-func (d *Database) GetDepositsByAddress(address common.Address, page PaginationParam) ([]DepositJSON, error) {
+// GetDepositsByAddress returns the list of Deposits indexed for the given
+// address paginated by the given params.
+func (d *Database) GetDepositsByAddress(address common.Address, page PaginationParam) (*PaginatedDeposits, error) {
 	const selectDepositsStatement = `
 	SELECT
 		deposits.guid, deposits.from_address, deposits.to_address,
@@ -634,17 +383,17 @@ func (d *Database) GetDepositsByAddress(address common.Address, page PaginationP
 
 		for rows.Next() {
 			var deposit DepositJSON
-			var l1_token Token
+			var l1Token Token
 			if err := rows.Scan(
 				&deposit.GUID, &deposit.FromAddress, &deposit.ToAddress,
 				&deposit.Amount, &deposit.TxHash, &deposit.Data,
-				&l1_token.Address, &deposit.L2Token,
-				&l1_token.Name, &l1_token.Symbol, &l1_token.Decimals,
+				&l1Token.Address, &deposit.L2Token,
+				&l1Token.Name, &l1Token.Symbol, &l1Token.Decimals,
 				&deposit.BlockNumber, &deposit.BlockTimestamp,
 			); err != nil {
 				return err
 			}
-			deposit.L1Token = &l1_token
+			deposit.L1Token = &l1Token
 			deposits = append(deposits, deposit)
 		}
 
@@ -654,9 +403,43 @@ func (d *Database) GetDepositsByAddress(address common.Address, page PaginationP
 	if err != nil {
 		return nil, err
 	}
-	return deposits, nil
+
+	const selectDepositCountStatement = `
+	SELECT
+		count(*)
+	FROM deposits
+		INNER JOIN l1_blocks ON deposits.block_hash=l1_blocks.hash
+		INNER JOIN l1_tokens ON deposits.l1_token=l1_tokens.address
+	WHERE deposits.from_address = $1;
+	`
+
+	var count uint64
+
+	err = txn(d.db, func(tx *sql.Tx) error {
+		queryStmt, err := tx.Prepare(selectDepositCountStatement)
+		if err != nil {
+			return err
+		}
+
+		row := queryStmt.QueryRow(address.String())
+		if err != nil {
+			return err
+		}
+
+		row.Scan(&count)
+		return nil
+	})
+
+	page.Total = count
+
+	return &PaginatedDeposits{
+		&page,
+		deposits,
+	}, nil
 }
 
+// GetWithdrawalBatch returns the StateBatch corresponding to the given
+// withdrawal transaction hash.
 func (d *Database) GetWithdrawalBatch(hash l2common.Hash) (*StateBatchJSON, error) {
 	const selectWithdrawalBatchStatement = `
 	SELECT
@@ -685,11 +468,11 @@ func (d *Database) GetWithdrawalBatch(hash l2common.Hash) (*StateBatchJSON, erro
 			return row.Err()
 		}
 
-		var index, size, prev_total, block_number, block_timestamp uint64
-		var root, block_hash string
-		var extra_data []byte
-		err = row.Scan(&index, &root, &size, &prev_total, &extra_data, &block_hash,
-			&block_number, &block_timestamp)
+		var index, size, prevTotal, blockNumber, blockTimestamp uint64
+		var root, blockHash string
+		var extraData []byte
+		err = row.Scan(&index, &root, &size, &prevTotal, &extraData, &blockHash,
+			&blockNumber, &blockTimestamp)
 		if err != nil {
 			if errors.Is(err, sql.ErrNoRows) {
 				batch = nil
@@ -702,11 +485,11 @@ func (d *Database) GetWithdrawalBatch(hash l2common.Hash) (*StateBatchJSON, erro
 			Index:          index,
 			Root:           root,
 			Size:           size,
-			PrevTotal:      prev_total,
-			ExtraData:      extra_data,
-			BlockHash:      block_hash,
-			BlockNumber:    block_number,
-			BlockTimestamp: block_timestamp,
+			PrevTotal:      prevTotal,
+			ExtraData:      extraData,
+			BlockHash:      blockHash,
+			BlockNumber:    blockNumber,
+			BlockTimestamp: blockTimestamp,
 		}
 
 		return nil
@@ -718,7 +501,9 @@ func (d *Database) GetWithdrawalBatch(hash l2common.Hash) (*StateBatchJSON, erro
 	return batch, nil
 }
 
-func (d *Database) GetWithdrawalsByAddress(address l2common.Address, page PaginationParam) ([]WithdrawalJSON, error) {
+// GetWithdrawalsByAddress returns the list of Withdrawals indexed for the given
+// address paginated by the given params.
+func (d *Database) GetWithdrawalsByAddress(address l2common.Address, page PaginationParam) (*PaginatedWithdrawals, error) {
 	const selectWithdrawalsStatement = `
 	SELECT
 	    withdrawals.guid, withdrawals.from_address, withdrawals.to_address,
@@ -746,17 +531,17 @@ func (d *Database) GetWithdrawalsByAddress(address l2common.Address, page Pagina
 
 		for rows.Next() {
 			var withdrawal WithdrawalJSON
-			var l2_token Token
+			var l2Token Token
 			if err := rows.Scan(
 				&withdrawal.GUID, &withdrawal.FromAddress, &withdrawal.ToAddress,
 				&withdrawal.Amount, &withdrawal.TxHash, &withdrawal.Data,
-				&withdrawal.L1Token, &l2_token.Address,
-				&l2_token.Name, &l2_token.Symbol, &l2_token.Decimals,
+				&withdrawal.L1Token, &l2Token.Address,
+				&l2Token.Name, &l2Token.Symbol, &l2Token.Decimals,
 				&withdrawal.BlockNumber, &withdrawal.BlockTimestamp,
 			); err != nil {
 				return err
 			}
-			withdrawal.L2Token = &l2_token
+			withdrawal.L2Token = &l2Token
 			withdrawals = append(withdrawals, withdrawal)
 		}
 
@@ -772,19 +557,41 @@ func (d *Database) GetWithdrawalsByAddress(address l2common.Address, page Pagina
 		withdrawals[i].Batch = batch
 	}
 
-	return withdrawals, nil
+	const selectWithdrawalCountStatement = `
+	SELECT
+		count(*)
+	FROM withdrawals
+		INNER JOIN l2_blocks ON withdrawals.block_hash=l2_blocks.hash
+		INNER JOIN l2_tokens ON withdrawals.l2_token=l2_tokens.address
+	WHERE withdrawals.from_address = $1;
+	`
+
+	var count uint64
+
+	err = txn(d.db, func(tx *sql.Tx) error {
+		queryStmt, err := tx.Prepare(selectWithdrawalCountStatement)
+		if err != nil {
+			return err
+		}
+
+		row := queryStmt.QueryRow(address.String())
+		if err != nil {
+			return err
+		}
+
+		row.Scan(&count)
+		return nil
+	})
+
+	page.Total = count
+
+	return &PaginatedWithdrawals{
+		&page,
+		withdrawals,
+	}, nil
 }
 
-type L1BlockLocator struct {
-	Number uint64      `json:"number"`
-	Hash   common.Hash `json:"hash"`
-}
-
-type L2BlockLocator struct {
-	Number uint64        `json:"number"`
-	Hash   l2common.Hash `json:"hash"`
-}
-
+// GetHighestL1Block returns the highest known L1 block.
 func (d *Database) GetHighestL1Block() (*L1BlockLocator, error) {
 	const selectHighestBlockStatement = `
 	SELECT number, hash FROM l1_blocks ORDER BY number DESC LIMIT 1
@@ -827,6 +634,7 @@ func (d *Database) GetHighestL1Block() (*L1BlockLocator, error) {
 	return highestBlock, nil
 }
 
+// GetHighestL2Block returns the highest known L2 block.
 func (d *Database) GetHighestL2Block() (*L2BlockLocator, error) {
 	const selectHighestBlockStatement = `
 	SELECT number, hash FROM l2_blocks ORDER BY number DESC LIMIT 1
@@ -869,6 +677,7 @@ func (d *Database) GetHighestL2Block() (*L2BlockLocator, error) {
 	return highestBlock, nil
 }
 
+// GetIndexedL1BlockByHash returns the L1 block by it's hash.
 func (d *Database) GetIndexedL1BlockByHash(hash common.Hash) (*IndexedL1Block, error) {
 	const selectBlockByHashStatement = `
 	SELECT
@@ -916,81 +725,9 @@ func (d *Database) GetIndexedL1BlockByHash(hash common.Hash) (*IndexedL1Block, e
 	}
 
 	return block, nil
-
 }
 
-func (d *Database) GetEventsByBlockHash(hash common.Hash) ([]TxnEnqueuedEvent, error) {
-	const selectEventsByBlockHashStatement = `
-	SELECT
-		b.number, b.timestamp,
-		d.tx_hash, d.data
-	FROM
-		blocks AS b,
-		deposits AS d
-	WHERE b.hash = d.block_hash AND b.hash = $1
-	`
-
-	var events []TxnEnqueuedEvent
-	err := txn(d.db, func(tx *sql.Tx) error {
-		queryStmt, err := tx.Prepare(selectEventsByBlockHashStatement)
-		if err != nil {
-			return err
-		}
-
-		rows, err := queryStmt.Query(hash.String())
-		if err != nil {
-			return err
-		}
-
-		for rows.Next() {
-			event, err := scanTxnEnqueuedEvent(rows)
-			if err != nil {
-				return err
-			}
-
-			events = append(events, event)
-		}
-
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return events, nil
-}
-
-func scanTxnEnqueuedEvent(rows *sql.Rows) (TxnEnqueuedEvent, error) {
-	var number uint64
-	var timestamp uint64
-	var txHash string
-	var data []byte
-	err := rows.Scan(
-		&number,
-		&timestamp,
-		&txHash,
-		&data,
-	)
-	if err != nil {
-		return TxnEnqueuedEvent{}, err
-	}
-
-	return TxnEnqueuedEvent{
-		BlockNumber: number,
-		Timestamp:   timestamp,
-		TxHash:      common.HexToHash(txHash),
-		Data:        data,
-	}, nil
-}
-
-func (d *Database) Close() error {
-	return d.db.Close()
-}
-
-func (d *Database) Config() string {
-	return d.config
-}
-
+// NewDatabase returns the database for the given connection string.
 func NewDatabase(config string) (*Database, error) {
 	db, err := sql.Open("postgres", config)
 	if err != nil {
@@ -1013,30 +750,4 @@ func NewDatabase(config string) (*Database, error) {
 		db:     db,
 		config: config,
 	}, nil
-}
-
-func txn(db *sql.DB, apply func(*sql.Tx) error) error {
-	tx, err := db.Begin()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if p := recover(); p != nil {
-			tx.Rollback()
-			panic(p)
-		}
-	}()
-
-	err = apply(tx)
-	if err != nil {
-		// Don't swallow application error
-		_ = tx.Rollback()
-		return err
-	}
-
-	return tx.Commit()
-}
-
-func NewGUID() string {
-	return uuid.New().String()
 }
